+++ src/Dak.RemoveWith.pas
@@ -10,42 +10,8 @@ function RunRemoveWithCommand(const aOptions: TAppOptions): Integer;
 implementation
 
 uses
-  System.IOUtils, System.JSON, System.SysUtils,
-  Dak.ExitCodes, Dak.Utils;
-
-function RemoveWithModeToText(const aMode: TRemoveWithMode): string;
-begin
-  case aMode of
-    TRemoveWithMode.rwmScan:
-      Result := 'scan';
-    TRemoveWithMode.rwmApply:
-      Result := 'apply';
-  else
-    Result := 'plan';
-  end;
-end;
-
-function RemoveWithFormatToText(const aFormat: TRemoveWithFormat): string;
-begin
-  if aFormat = TRemoveWithFormat.rwfText then
-    Result := 'text'
-  else
-    Result := 'json';
-end;
-
-function RemoveWithTargetKindToText(const aKind: TRemoveWithTargetKind): string;
-begin
-  case aKind of
-    TRemoveWithTargetKind.rwtUnit:
-      Result := 'unit';
-    TRemoveWithTargetKind.rwtDir:
-      Result := 'dir';
-    TRemoveWithTargetKind.rwtAll:
-      Result := 'all';
-  else
-    Result := 'none';
-  end;
-end;
+  System.IOUtils, System.SysUtils,
+  Dak.ExitCodes, Dak.RemoveWith.Output, Dak.Utils;
 
 function NewRemoveWithRunId: string;
 var
@@ -59,76 +25,6 @@ begin
   Result := StringReplace(Result, '}', '', [rfReplaceAll]);
   Result := StringReplace(Result, '-', '', [rfReplaceAll]);
   Result := LowerCase(Result);
-end;
-
-function BuildRemoveWithJsonReport(const aOptions: TAppOptions; const aProjectPath, aWorkspaceRoot, aUnitPath,
-  aDirPath: string): string;
-var
-  lProjectJson: TJSONObject;
-  lRoot: TJSONObject;
-  lSummary: TJSONObject;
-  lTargets: TJSONObject;
-  lWorkspace: TJSONObject;
-begin
-  lRoot := TJSONObject.Create;
-  try
-    lRoot.AddPair('operation', 'remove-with');
-    lRoot.AddPair('mode', RemoveWithModeToText(aOptions.fRemoveWithMode));
-    lRoot.AddPair('format', RemoveWithFormatToText(aOptions.fRemoveWithFormat));
-
-    lProjectJson := TJSONObject.Create;
-    lProjectJson.AddPair('path', aProjectPath);
-    lProjectJson.AddPair('name', TPath.GetFileNameWithoutExtension(aProjectPath));
-    lProjectJson.AddPair('dir', TPath.GetDirectoryName(aProjectPath));
-    lRoot.AddPair('project', lProjectJson);
-
-    lTargets := TJSONObject.Create;
-    lTargets.AddPair('kind', RemoveWithTargetKindToText(aOptions.fRemoveWithTargetKind));
-    lTargets.AddPair('unit', aUnitPath);
-    lTargets.AddPair('dir', aDirPath);
-    lTargets.AddPair('all', TJSONBool.Create(aOptions.fRemoveWithAll));
-    lRoot.AddPair('targets', lTargets);
-
-    lWorkspace := TJSONObject.Create;
-    lWorkspace.AddPair('root', aWorkspaceRoot);
-    lWorkspace.AddPair('reports', TPath.Combine(aWorkspaceRoot, 'reports'));
-    lWorkspace.AddPair('tmp', TPath.Combine(aWorkspaceRoot, 'tmp'));
-    lRoot.AddPair('workspace', lWorkspace);
-
-    lSummary := TJSONObject.Create;
-    lSummary.AddPair('filesScanned', TJSONNumber.Create(0));
-    lSummary.AddPair('withStatements', TJSONNumber.Create(0));
-    lSummary.AddPair('plannedEdits', TJSONNumber.Create(0));
-    lSummary.AddPair('appliedEdits', TJSONNumber.Create(0));
-    lSummary.AddPair('skipped', TJSONNumber.Create(0));
-    lSummary.AddPair('failed', TJSONNumber.Create(0));
-    lSummary.AddPair('rolledBack', TJSONNumber.Create(0));
-    lRoot.AddPair('summary', lSummary);
-
-    lRoot.AddPair('warnings', TJSONArray.Create);
-    Result := lRoot.ToJSON;
-  finally
-    lRoot.Free;
-  end;
-end;
-
-function BuildRemoveWithTextReport(const aOptions: TAppOptions; const aProjectPath, aWorkspaceRoot, aUnitPath,
-  aDirPath: string): string;
-var
-  lTargetValue: string;
-begin
-  lTargetValue := aUnitPath;
-  if aOptions.fRemoveWithTargetKind = TRemoveWithTargetKind.rwtDir then
-    lTargetValue := aDirPath
-  else if aOptions.fRemoveWithTargetKind = TRemoveWithTargetKind.rwtAll then
-    lTargetValue := '<all project units>';
-
-  Result := 'operation=remove-with' + sLineBreak +
-    'mode=' + RemoveWithModeToText(aOptions.fRemoveWithMode) + sLineBreak +
-    'project=' + aProjectPath + sLineBreak +
-    'target=' + RemoveWithTargetKindToText(aOptions.fRemoveWithTargetKind) + ':' + lTargetValue + sLineBreak +
-    'workspace=' + aWorkspaceRoot + sLineBreak +
-    'withStatements=0';
 end;
 
 procedure WriteRemoveWithOutput(const aOptions: TAppOptions; const aOutputText: string);
@@ -205,9 +101,9 @@ begin
   TDirectory.CreateDirectory(lWorkspaceRoot);
 
   if aOptions.fRemoveWithFormat = TRemoveWithFormat.rwfText then
-    lOutputText := BuildRemoveWithTextReport(aOptions, lProjectPath, lWorkspaceRoot, lUnitPath, lDirPath)
+    lOutputText := BuildRemoveWithTextReport(aOptions, lProjectPath, lWorkspaceRoot, lRunId, lUnitPath, lDirPath)
   else
-    lOutputText := BuildRemoveWithJsonReport(aOptions, lProjectPath, lWorkspaceRoot, lUnitPath, lDirPath);
+    lOutputText := BuildRemoveWithJsonReport(aOptions, lProjectPath, lWorkspaceRoot, lRunId, lUnitPath, lDirPath);
   WriteRemoveWithOutput(aOptions, lOutputText);
   Result := cExitSuccess;
 end;